Refactor peagen models into modular structure
