+++ pkgs/standards/peagen/peagen/models/__init__.py
@@ -1,9 +1,11 @@
 from __future__ import annotations
 
-from peagen.models.task_run import Base, TaskRun, TaskRunDep
-from peagen.models.secret import Secret
+from peagen.models.base import Base
+from peagen.models.task import TaskRun, TaskRunDep
+from peagen.models.config import Secret
 from peagen.models.abuse import AbuseRecord
-from peagen.models.schemas import Pool, Role, Status, Task, User
+from peagen.models.schemas import Pool, Role, Status, Task
+from peagen.models.tenant import Tenant, User
 
 __all__ = [
     "Role",
@@ -11,6 +13,7 @@ __all__ = [
     "Task",
     "Pool",
     "User",
+    "Tenant",
     "Base",
     "TaskRun",
     "Secret",
