--- pkgs/standards/peagen/peagen/models/core_models.py
@@ -1,111 +1,0 @@
-import datetime as dt
-import uuid
-
-from sqlalchemy import ForeignKey, String, UniqueConstraint, TIMESTAMP, JSON, Text
-from sqlalchemy.dialects.postgresql import UUID
-from sqlalchemy.orm import Mapped, mapped_column
-
-from .task_run import Base
-
-
-class Tenant(Base):
-    __tablename__ = "tenants"
-
-    id: Mapped[uuid.UUID] = mapped_column(
-        UUID(as_uuid=True), primary_key=True, default=uuid.uuid4
-    )
-    name: Mapped[str] = mapped_column(String, nullable=False, unique=True)
-    created_at: Mapped[dt.datetime] = mapped_column(
-        TIMESTAMP(timezone=True), default=dt.datetime.utcnow
-    )
-    updated_at: Mapped[dt.datetime] = mapped_column(
-        TIMESTAMP(timezone=True),
-        default=dt.datetime.utcnow,
-        onupdate=dt.datetime.utcnow,
-    )
-
-
-class _TenantBase:
-    tenant_id: Mapped[uuid.UUID] = mapped_column(
-        ForeignKey("tenants.id"), nullable=False
-    )
-    name: Mapped[str] = mapped_column(String, nullable=False)
-    created_at: Mapped[dt.datetime] = mapped_column(
-        TIMESTAMP(timezone=True), default=dt.datetime.utcnow
-    )
-    updated_at: Mapped[dt.datetime] = mapped_column(
-        TIMESTAMP(timezone=True),
-        default=dt.datetime.utcnow,
-        onupdate=dt.datetime.utcnow,
-    )
-
-
-class User(Base, _TenantBase):
-    __tablename__ = "users"
-
-    id: Mapped[uuid.UUID] = mapped_column(
-        UUID(as_uuid=True), primary_key=True, default=uuid.uuid4
-    )
-
-    __table_args__ = (
-        UniqueConstraint("tenant_id", "name", name="uq_users_tenant_name"),
-    )
-
-
-class PublicKey(Base, _TenantBase):
-    __tablename__ = "public_keys"
-
-    id: Mapped[uuid.UUID] = mapped_column(
-        UUID(as_uuid=True), primary_key=True, default=uuid.uuid4
-    )
-    user_id: Mapped[uuid.UUID] = mapped_column(ForeignKey("users.id"), nullable=False)
-    key: Mapped[str] = mapped_column(Text, nullable=False)
-
-    __table_args__ = (
-        UniqueConstraint("tenant_id", "name", name="uq_public_keys_tenant_name"),
-    )
-
-
-class Task(Base, _TenantBase):
-    __tablename__ = "tasks"
-
-    id: Mapped[uuid.UUID] = mapped_column(
-        UUID(as_uuid=True), primary_key=True, default=uuid.uuid4
-    )
-    payload: Mapped[dict | None] = mapped_column(JSON)
-
-    __table_args__ = (
-        UniqueConstraint("tenant_id", "name", name="uq_tasks_tenant_name"),
-    )
-
-
-class Artifact(Base, _TenantBase):
-    __tablename__ = "artifacts"
-
-    id: Mapped[uuid.UUID] = mapped_column(
-        UUID(as_uuid=True), primary_key=True, default=uuid.uuid4
-    )
-    task_id: Mapped[uuid.UUID | None] = mapped_column(
-        ForeignKey("tasks.id"), nullable=True
-    )
-    uri: Mapped[str] = mapped_column(String, nullable=False)
-
-    __table_args__ = (
-        UniqueConstraint("tenant_id", "name", name="uq_artifacts_tenant_name"),
-    )
-
-
-class DeployKey(Base, _TenantBase):
-    __tablename__ = "deploy_keys"
-
-    id: Mapped[uuid.UUID] = mapped_column(
-        UUID(as_uuid=True), primary_key=True, default=uuid.uuid4
-    )
-    public_key_id: Mapped[uuid.UUID] = mapped_column(
-        ForeignKey("public_keys.id"), nullable=False
-    )
-    key: Mapped[str] = mapped_column(Text, nullable=False)
-
-    __table_args__ = (
-        UniqueConstraint("tenant_id", "name", name="uq_deploy_keys_tenant_name"),
-    )
