+++ pkgs/standards/peagen/peagen/models/result/__init__.py
@@ -1,0 +1,6 @@
+from __future__ import annotations
+
+from .eval_result import EvalResult
+from .analysis_result import AnalysisResult
+
+__all__ = ["EvalResult", "AnalysisResult"]
