+++ pkgs/standards/peagen/peagen/models/tenant/__init__.py
@@ -1,0 +1,6 @@
+from __future__ import annotations
+
+from .tenant import Tenant
+from .user import User
+
+__all__ = ["Tenant", "User"]
