+++ pkgs/standards/peagen/peagen/models/tenant/tenant_user_association.py
@@ -1,0 +1,22 @@
+from __future__ import annotations
+
+import uuid
+
+from sqlalchemy import ForeignKey
+from sqlalchemy.dialects.postgresql import UUID
+from sqlalchemy.orm import Mapped, mapped_column
+
+from peagen.models.base import Base
+
+
+class TenantUserAssociation(Base):
+    """Link table associating tenants with users."""
+
+    __tablename__ = "tenant_user_associations"
+
+    tenant_id: Mapped[uuid.UUID] = mapped_column(
+        UUID(as_uuid=True), ForeignKey("tenants.id"), primary_key=True
+    )
+    user_id: Mapped[uuid.UUID] = mapped_column(
+        UUID(as_uuid=True), ForeignKey("users.id"), primary_key=True
+    )
