+++ pkgs/standards/peagen/peagen/models/tenant/user.py
@@ -1,0 +1,24 @@
+from __future__ import annotations
+
+import uuid
+
+from sqlalchemy import ForeignKey, String, UniqueConstraint
+from sqlalchemy.dialects.postgresql import UUID
+from sqlalchemy.orm import Mapped, mapped_column
+
+from peagen.models.base import Base, UUIDMixin, TimestampMixin
+
+
+class User(Base, UUIDMixin, TimestampMixin):
+    """User belonging to a tenant."""
+
+    __tablename__ = "users"
+
+    tenant_id: Mapped[uuid.UUID] = mapped_column(
+        UUID(as_uuid=True), ForeignKey("tenants.id"), nullable=False
+    )
+    name: Mapped[str] = mapped_column(String, nullable=False)
+
+    __table_args__ = (
+        UniqueConstraint("tenant_id", "name", name="uq_users_tenant_name"),
+    )
